requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
pycpfcnpj==1.5.1
pytz==2021.1
sqlparse==0.4.1
psycopg2